requirements.txt
Django==1.6
psycopg2
solrpy
django-tastypie
django-grappelli==2.5.2
